MWPW-145148 Inline tree.js into fragment.js (#2052)
* MWPW-145148 Inline tree.js into fragment.js

No code change, just a move of code in tree.js into fragment.js

* Move tree test
