+++ libs/blocks/fragment/fragment.js
@@ -1,5 +1,5 @@
+/* eslint-disable max-classes-per-file */
 import { createTag, getConfig, loadArea, localizeLink } from '../../utils/utils.js';
-import Tree from '../../utils/tree.js';
 
 const fragMap = {};
 
@@ -20,6 +20,7 @@ const updateFragMap = (fragment, a, href) => {
   if (!fragLinks.length) return;
 
   if (document.body.contains(a)) { // is fragment on page (not nested)
+    // eslint-disable-next-line no-use-before-define
     fragMap[href] = new Tree(href);
     fragLinks.forEach((link) => fragMap[href].insert(href, localizeLink(removeHash(link.href))));
   } else {
@@ -130,5 +131,61 @@ export default async function init(a) {
   } else {
     a.parentElement.replaceChild(fragment, a);
     await loadArea(fragment);
+  }
+}
+
+class Node {
+  constructor(key, value = key, parent = null) {
+    this.key = key;
+    this.value = value;
+    this.parent = parent;
+    this.children = [];
+  }
+
+  get isLeaf() {
+    return this.children.length === 0;
+  }
+}
+
+export class Tree {
+  constructor(key, value = key) {
+    this.root = new Node(key, value);
+  }
+
+  * traverse(node = this.root) {
+    yield node;
+    if (node.children.length) {
+      for (const child of node.children) {
+        yield* this.traverse(child);
+      }
+    }
+  }
+
+  insert(parentNodeKey, key, value = key) {
+    for (const node of this.traverse()) {
+      if (node.key === parentNodeKey) {
+        node.children.push(new Node(key, value, node));
+        return true;
+      }
+    }
+    return false;
+  }
+
+  remove(key) {
+    for (const node of this.traverse()) {
+      const filtered = node.children.filter((c) => c.key !== key);
+      if (filtered.length !== node.children.length) {
+        node.children = filtered;
+        return true;
+      }
+    }
+    return false;
+  }
+
+  find(key) {
+    for (const node of this.traverse()) {
+      if (node.key === key) return node;
+    }
+    return undefined;
   }
 }
