+++ test/blocks/fragment/tree.test.js
@@ -1,5 +1,5 @@
 import { expect } from '@esm-bundle/chai';
-import Tree from '../../libs/utils/tree.js';
+import { Tree } from '../../../libs/blocks/fragment/fragment.js';
 
 describe('Tree Data Struct', () => {
   it('should create a tree with root node', () => {
